requirements.txt
apache-airflow==2.3.0
apache-airflow-providers-snowflake==4.0.4
google-cloud-storage==1.43.0
apache-airflow.providers.postgres==2.2.0
protobuf==3.20.1
pinecone-client>=2.2.2